refactor: extract Win32 window placement runtime from main
Made-with: Cursor
